# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Documentación de API ",
    "version": "1.0.0",
    "title": "API BIBLIOTECA UNIVALLE",
    "contact": {
      "email": "vqg2021079@est.univalle.edu"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost",
  "basePath": "/univalle/v1",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/students": {
      "get": {
        "tags": [
          "Student"
        ],
        "summary": "Listado de estudiantes",
        "description": "",
        "operationId": "listStudent",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "List"
          }
        }
      },
      "post": {
        "tags": [
          "Student"
        ],
        "summary": "Add a new student in the register",
        "description": "",
        "operationId": "addStudent",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Student object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Student"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Student created"
          }
        }
      },
      "put": {
        "tags": [
          "Student"
        ],
        "summary": "Update an existing pet",
        "description": "",
        "operationId": "updatePet",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Pet object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Student"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/student/{id}": {
      "get": {
        "tags": [
          "Student"
        ],
        "summary": "Find student by ID",
        "description": "Returns a single student",
        "operationId": "getStudentbyid",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of student to return",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Student"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Student not found"
          }
        }
      },
      "patch": {
        "tags": [
          "Student"
        ],
        "summary": "Updates a student in the register with form data",
        "description": "",
        "operationId": "updateStudentWithForm",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of pet that needs to be updated",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "Updated name of the pet",
            "required": false,
            "type": "string"
          },
          {
            "name": "status",
            "in": "formData",
            "description": "Updated status of the pet",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Invalid input"
          }
        }
      },
      "delete": {
        "tags": [
          "Student"
        ],
        "summary": "Deletes a Student",
        "description": "",
        "operationId": "deleteStudent",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Pet id to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          }
        }
      }
    },
...
    "/loan": {
      "get": {
        "tags": [
          "loan"
        ],
        "summary": "Listado de prestamos",
        "description": "",
        "operationId": "listLoan",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "List"
          }
        }
      },
      "post": {
        "tags": [
          "loan"
        ],
        "summary": "Add a new loan in the register",
        "description": "",
        "operationId": "addLoan",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "loan object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Loan"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Loan created"
          }
        }
      },
      "put": {
        "tags": [
          "loan"
        ],
        "summary": "Update an existing loan",
        "description": "",
        "operationId": "updateLoan",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "LOan object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Loan"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/loan/{id}": {
      "get": {
        "tags": [
          "loan"
        ],
        "summary": "Find loan by ID",
        "description": "Returns a single loan",
        "operationId": "getLoanbyid",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of loan return",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Loan"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Loan not found"
          }
        }
      },
      "patch": {
        "tags": [
          "loan"
        ],
        "summary": "Updates a loan in the register with form data",
        "description": "",
        "operationId": "updateLoantWithForm",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of pet that needs to be updated",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "Updated name of the loan",
            "required": false,
            "type": "string"
          },
          {
            "name": "status",
            "in": "formData",
            "description": "Updated status of the loan",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Invalid input"
          }
        }
      },
      "delete": {
        "tags": [
          "loan"
        ],
        "summary": "Deletes a Loan",
        "description": "",
        "operationId": "deleteLoan",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Loan id to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Loan not found"
          }
        }
      }
    },
    "/reserve": {
      "get": {
        "tags": [
          "reserve"
        ],
        "summary": "Listado de reservas",
        "description": "",
        "operationId": "listReserve",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "List"
          }
        }
      },
      "post": {
        "tags": [
          "reserve"
        ],
        "summary": "Add a new reserve in the register",
        "description": "",
        "operationId": "addReserve",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "reserve object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/reserve"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "reserve created"
          }
        }
      },
      "put": {
        "tags": [
          "reserve"
        ],
        "summary": "Update an existing reserve",
        "description": "",
        "operationId": "updatereserve",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "reserve object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/reserve"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "reserve not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/reserve/{id}": {
      "get": {
        "tags": [
          "reserve"
        ],
        "summary": "Find loan by ID",
        "description": "Returns a single loan",
        "operationId": "getreservebyid",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of loan return",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Loan"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Loan not found"
          }
        }
      },
      "patch": {
        "tags": [
          "reserve"
        ],
        "summary": "Updates a student in the register with form data",
        "description": "",
        "operationId": "updateReservetWithForm",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of pet that needs to be updated",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "Updated name of the pet",
            "required": false,
            "type": "string"
          },
          {
            "name": "status",
            "in": "formData",
            "description": "Updated status of the pet",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Invalid input"
          }
        }
      },
      "delete": {
        "tags": [
          "reserve"
        ],
        "summary": "Deletes a Student",
        "description": "",
        "operationId": "deleteReserve",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Pet id to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Student": {
      "type": "object",
      "required": [
        "name",
        "image",
        "age",
        "phone"
      ],
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "name": {
          "type": "string",
          "example": "Juan"
        },
        "image": {
          "type": "string"
        },
        "age": {
          "type": "integer"
        },
        "phone": {
          "type": "string",
          "description": "phone's student"
        }
      }
    },
...
    "Loan": {
      "type": "object",
      "required": [
        "id_book",
        "id_student",
        "date"
      ],
      "properties": {
        "id_book": {
          "type": "string",
          "format": "uuid"
        },
        "id_student": {
          "type": "string",
          "example": "CVG-0503"
        },
        "date": {
          "type": "string"
        }
      }
    },
    "reserve": {
      "type": "object",
      "required": [
        "id_book",
        "id_student",
        "date"
      ],
      "properties": {
        "id_book": {
          "type": "string",
          "format": "uuid"
        },
        "id_student": {
          "type": "string",
          "example": "CVG-0503"
        },
        "date": {
          "type": "string"
        }
      }
    },
...
  }
}